requirements.txt
Django==1.9.6
django-filter==0.13.0
djangorestframework==3.4.3
Markdown==2.6.6
pkg-resources==0.0.0
xlrd==0.9.4
